# doc/swagger.json
{"swagger":"2.0","info":{"title":"My book API Docs","version":"1.0.0","description":"API documentation for my book project"},"paths":{"/admins":{"post":{"tags":["Admin"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/CreateAdmin"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/CreateAdmin"}}}}},"get":{"tags":["Admin"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetAdmins"}}}}}},"/admins/{id}":{"delete":{"tags":["Admin"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Admin UUID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Boolean","items":{"$ref":"#/definitions/DeleteAdmin"}}}}},"get":{"tags":["Admin"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Admin UUID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/GetAdminById"}}}}},"patch":{"tags":["Admin"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Admin UUID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UpdateAdmin"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/UpdateAdmin"}}}}}},"/admins/profile":{"get":{"tags":["Admin"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/GetAdminProfile"}}}}}},"/admins/{id}/status":{"patch":{"tags":["Admin"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Admin UUID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UpdateAdminStatus"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/UpdateAdminStatus"}}}}}},"/admin/auth/login":{"post":{"tags":["Admin_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/AdminLogin"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/AdminLogin"}}}}}},"/admin/auth/logout":{"post":{"tags":["Admin_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/null"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Boolean","items":{"$ref":"#/definitions/AdminLogout"}}}}}},"/upload":{"post":{"tags":["File_Management"],"summary":"Upload a file to the server. Supports various file types including images, documents, and other allowed formats.","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Upload a file to the server. Supports various file types including images, documents, and other allowed formats.","schema":{"$ref":"#/definitions/UploadFile"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/UploadFile"}}}}}},"/admin/roles":{"post":{"tags":["Role_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/CreateRoleAdmin"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/CreateRoleAdmin"}}}}},"get":{"tags":["Role_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/ListRoles"}}}}}},"/admin/roles/{id}":{"delete":{"tags":["Role_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Role ID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Boolean","items":{"$ref":"#/definitions/DeleteRole"}}}}},"get":{"tags":["Role_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Role ID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/GetRole"}}}}},"patch":{"tags":["Role_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Role ID"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/UpdateRole"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/UpdateRole"}}}}}},"/roles":{"post":{"tags":["Roles"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/CreateRoleCustomer"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/CreateRoleCustomer"}}}}}},"/api/v1/healths/ping":{"get":{"tags":["System_Health"],"summary":"Check the health status of the API service","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/Get_status_health_check"}}}}}},"/auth/refresh":{"post":{"tags":["User_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/RefreshToken"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/RefreshToken"}}}}}},"/auth/register":{"post":{"tags":["User_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/RegisterUser"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/RegisterUser"}}}}}},"/auth/verify-account/resend":{"post":{"tags":["User_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/ResendVerificationOTP"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/ResendVerificationOTP"}}}}}},"/auth/login/socials/{social}":{"get":{"tags":["User_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"social","in":"path","required":true,"type":"string","description":"Social provider (e.g., &quot;google&quot;, &quot;facebook&quot;)"}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/SocialLogin"}}}}}},"/auth/login":{"post":{"tags":["User_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/UserLogin"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/UserLogin"}}}}}},"/auth/logout":{"post":{"tags":["User_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/null"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Boolean","items":{"$ref":"#/definitions/UserLogout"}}}}}},"/auth/verify-account":{"post":{"tags":["User_Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/VerifyAccount"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/VerifyAccount"}}}}}},"/admin/users/{id}":{"delete":{"tags":["User_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"User's unique ID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Boolean","items":{"$ref":"#/definitions/DeleteUser"}}}}},"get":{"tags":["User_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"User's unique ID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/GetUser"}}}}}},"/admin/users":{"get":{"tags":["User_Management"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/ListUsers"}}}}}},"/users/profile":{"get":{"tags":["User_Profile"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/GetProfile"}}}}},"patch":{"tags":["User_Profile"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/UpdateProfile"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/UpdateProfile"}}}}}}},"definitions":{"CreateAdmin":{"properties":{"id":{"type":"string","description":"Created admin ID"},"email":{"type":"string","description":"Admin email"}},"required":["id","email"]},"GetAdmins":{"properties":{"admins":{"type":"array","description":"List of admin users","items":{"type":"Object"}}},"required":["admins"]},"admins":{"properties":{"id":{"type":"string","description":"Admin ID"},"email":{"type":"string","description":"Admin email"}},"required":["id","email"]},"DeleteAdmin":{"properties":{"id":{"type":"string","description":"Admin UUID"},"success":{"type":"boolean","description":"Deletion status"}},"required":["id","success"]},"GetAdminById":{"properties":{"id":{"type":"string","description":"Admin ID"},"email":{"type":"string","description":"Admin email"},"profile":{"type":"object","description":"Admin profile","$ref":"#/definitions/profile"}},"required":["id","email","profile"]},"UpdateAdmin":{"properties":{"id":{"type":"string","description":"Admin ID"},"email":{"type":"string","description":"Updated email"}},"required":["id","email"]},"GetAdminProfile":{"properties":{"id":{"type":"string","description":"Admin ID"},"email":{"type":"string","description":"Admin email"},"profile":{"type":"object","description":"Admin profile information","$ref":"#/definitions/profile"}},"required":["id","email","profile"]},"UpdateAdminStatus":{"properties":{"id":{"type":"string","description":"Admin ID"},"status":{"type":"string","description":"Updated status"}},"required":["id","status"]},"AdminLogin":{"properties":{"token":{"type":"string","description":"JWT access token"},"refreshToken":{"type":"string","description":"Refresh token"},"user":{"type":"object","description":"Admin user details","$ref":"#/definitions/user"}},"required":["token","refreshToken","user"]},"AdminLogout":{"properties":{"success":{"type":"boolean","description":"Logout status"}},"required":["success"]},"UploadFile":{"properties":{"id":{"type":"string","description":"Uploaded file ID"},"url":{"type":"string","description":"File access URL"},"name":{"type":"string","description":"Original filename"},"mimeType":{"type":"string","description":"File MIME type"},"size":{"type":"number","description":"File size in bytes"},"uploadedAt":{"type":"date","description":"Upload timestamp"}},"required":["id","url","name","mimeType","size","uploadedAt"]},"CreateRoleAdmin":{"properties":{"id":{"type":"string","description":"Created role ID"},"userId":{"type":"string","description":"Customer ID"},"items":{"type":"array","description":"Role items","items":{"type":"Object"}},"total":{"type":"number","description":"Role total"},"status":{"type":"string","description":"Role status"}},"required":["id","userId","items","total","status"]},"ListRoles":{"properties":{"roles":{"type":"array","description":"List of roles","items":{"type":"Object"}}},"required":["roles"]},"roles":{"properties":{"id":{"type":"string","description":"Role ID"},"userId":{"type":"string","description":"Customer ID"},"total":{"type":"number","description":"Role total amount"},"status":{"type":"string","description":"Role status"},"createdAt":{"type":"date","description":"Role creation date"}},"required":["id","userId","total","status","createdAt"]},"DeleteRole":{"properties":{"id":{"type":"string","description":"Role ID"},"success":{"type":"boolean","description":"Deletion status"}},"required":["id","success"]},"GetRole":{"properties":{"id":{"type":"string","description":"Role ID"},"userId":{"type":"string","description":"Customer ID"},"items":{"type":"array","description":"Role items","items":{"type":"Object"}},"total":{"type":"number","description":"Role total"},"status":{"type":"string","description":"Role status"},"createdAt":{"type":"date","description":"Creation date"}},"required":["id","userId","items","total","status","createdAt"]},"UpdateRole":{"properties":{"id":{"type":"string","description":"Role ID"},"status":{"type":"string","description":"Updated status"}},"required":["id","status"]},"CreateRoleCustomer":{"properties":{"id":{"type":"string","description":"Created role ID"},"items":{"type":"array","description":"Role items","items":{"type":"Object"}},"total":{"type":"number","description":"Role total"},"discount":{"type":"number","description":"Applied discount"},"status":{"type":"string","description":"Role status"}},"required":["id","items","total","discount","status"]},"Get_status_health_check":{"properties":{"status":{"type":"string","description":"Current health status"},"timestamp":{"type":"string","description":"Current server timestamp"},"info":{"type":"object","description":"Additional health information","$ref":"#/definitions/info"}},"required":["status","timestamp","info"]},"info":{"properties":{"version":{"type":"string","description":"API version"},"environment":{"type":"string","description":"Current environment"}},"required":["version","environment"]},"RefreshToken":{"properties":{"token":{"type":"string","description":"New JWT access token"},"refreshToken":{"type":"string","description":"New refresh token"}},"required":["token","refreshToken"]},"RegisterUser":{"properties":{"message":{"type":"string","description":"Registration success message"},"user":{"type":"object","description":"Created user details","$ref":"#/definitions/user"}},"required":["message","user"]},"ResendVerificationOTP":{"properties":{"message":{"type":"string","description":"OTP resend confirmation"}},"required":["message"]},"SocialLogin":{"properties":{"social":{"type":"string","description":"Social provider (e.g., &quot;google&quot;, &quot;facebook&quot;)"},"redirectUrl":{"type":"string","description":"Social login URL"}},"required":["social","redirectUrl"]},"UserLogin":{"properties":{"token":{"type":"string","description":"JWT access token"},"refreshToken":{"type":"string","description":"Refresh token"},"user":{"type":"object","description":"User details","$ref":"#/definitions/user"}},"required":["token","refreshToken","user"]},"UserLogout":{"properties":{"success":{"type":"boolean","description":"Logout status"}},"required":["success"]},"VerifyAccount":{"properties":{"message":{"type":"string","description":"Verification success message"},"user":{"type":"object","description":"Verified user details","$ref":"#/definitions/user"}},"required":["message","user"]},"DeleteUser":{"properties":{"id":{"type":"string","description":"User's unique ID"},"success":{"type":"boolean","description":"Operation status"},"message":{"type":"string","description":"Success message"}},"required":["id","success","message"]},"GetUser":{"properties":{"id":{"type":"string","description":"User ID"},"email":{"type":"string","description":"Email address"},"name":{"type":"string","description":"Full name"},"status":{"type":"string","description":"Account status"},"roles":{"type":"array","description":"User roles","items":{"type":"String"}},"createdAt":{"type":"date","description":"Registration date"},"lastLoginAt":{"type":"date","description":"Last login date"}},"required":["id","email","name","status","roles","createdAt","lastLoginAt"]},"ListUsers":{"properties":{"users":{"type":"array","description":"List of users","items":{"type":"Object"}},"pagination":{"type":"object","description":"Pagination details","$ref":"#/definitions/pagination"}},"required":["users","pagination"]},"users":{"properties":{"id":{"type":"string","description":"User ID"},"email":{"type":"string","description":"Email address"},"name":{"type":"string","description":"Full name"},"status":{"type":"string","description":"Account status"},"roles":{"type":"array","description":"User roles","items":{"type":"String"}},"createdAt":{"type":"date","description":"Registration date"}},"required":["id","email","name","status","roles","createdAt"]},"GetProfile":{"properties":{"id":{"type":"string","description":"User ID"},"email":{"type":"string","description":"Email address"},"name":{"type":"string","description":"Full name"},"preferences":{"type":"object","description":"User preferences","$ref":"#/definitions/preferences"},"createdAt":{"type":"date","description":"Registration date"}},"required":["id","email","name","preferences","createdAt"]},"UpdateProfile":{"properties":{"id":{"type":"string","description":"User ID"},"name":{"type":"string","description":"Updated name"},"preferences":{"type":"object","description":"Updated preferences","$ref":"#/definitions/preferences"}},"required":["id","name","preferences"]}}}